pyproject.toml
[tool.poetry]
name = "dolphin-bot"
version = "0.1.0"
description = "A simple bot for keeping track of beats to beat"
authors = ["realstealthninja <realstealthninja@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
disnake = "^2.9.1"
disnake-jishaku = "^2.6.5"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
aiosqlite = "^0.19.0"
python-dateutil = "^2.9.0.post0"

[tool.black]
line-length = 79

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
